backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
python-multipart==0.0.9
pydantic==2.7.3
pydantic-settings==2.3.4
supabase==2.5.3
openai==1.35.3
anthropic==0.29.0
pytesseract==0.3.13
Pillow==10.4.0
PyMuPDF==1.24.5
httpx==0.27.0
python-dotenv==1.0.1
numpy==1.26.4
